Refactor, add encryptionRepository.
